refactor: move StateFlowExtensions in dedicated package
